4.21* done

+++ osa4/bloglist/controllers/blogs.js
@@ -38,8 +38,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const requestingUser = await User.findById(decodedToken.id)
+  const deletableBlog = await Blog.findById(request.params.id)
+  if (requestingUser.id === deletableBlog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
